windowWidth from store
