+++ src/components/SimpleLayoutContainer.vue
@@ -7,6 +7,7 @@
 <script>
   import BaseSimpleContainer from "../mixins/BaseSimpleContainer"
   import { ref, provide, computed, onMounted, getCurrentInstance } from "vue"
+  import { useWindowWidth } from "@/store"
 
   export default {
     mixins: [BaseSimpleContainer],
@@ -28,26 +29,11 @@
     },
     setup() {
       const app = getCurrentInstance()
-      const breakpoints =
-        app.appContext.config.globalProperties.$SimpleVueGrid.breakpoints
+      // const breakpoints =
+      //   app.appContext.config.globalProperties.$SimpleVueGrid.breakpoints
 
-      const windowWidth = ref(window.innerWidth)
-      onMounted(() => {
-        window.onresize = () => {
-          windowWidth.value = window.innerWidth
-        }
-      })
-      const breakpoint = computed(() => {
-        let result
-        const Points = breakpoints
-        for (let point in Points) {
-          if (Points[point] <= windowWidth.value) {
-            result = point
-          }
-        }
+      const { windowWidth, breakpoint } = useWindowWidth()
 
-        return result
-      })
       provide("currentBreakpoint", breakpoint)
 
       return {
